refactor: move type_hint_utils out into own module
